fixed int conversion bug in cost function

+++ server/webapp/serializers/project_progsets.py
@@ -53,8 +53,8 @@ costcov_param_parser.add_arguments({
     'year': {'required': True, 'type': int},
     'saturationpercent_lower': {'required': True, 'type': float},
     'saturationpercent_upper': {'required': True, 'type': float},
-    'unitcost_lower': {'required': True, 'type': int},
-    'unitcost_upper': {'required': True, 'type': int},
+    'unitcost_lower': {'required': True, 'type': float},
+    'unitcost_upper': {'required': True, 'type': float},
 })
